Move most OpenGL texture allocation and upload code to a separate source file
