+++ src/image.go
@@ -9,7 +9,6 @@ import (
 	"io"
 	"math"
 	"os"
-	"runtime"
 	"unsafe"
 
 	gl "github.com/fyne-io/gl-js"
@@ -25,23 +24,6 @@ const (
 	TT_add1
 	TT_sub
 )
-
-type Texture struct {
-	handle gl.Texture
-}
-
-func newTexture() (t *Texture) {
-	t = &Texture{gl.CreateTexture()}
-	runtime.SetFinalizer(t, (*Texture).finalizer)
-	return
-}
-func (t *Texture) finalizer() {
-	if t.handle.IsValid() {
-		sys.mainThreadTask <- func() {
-			gl.DeleteTexture(t.handle)
-		}
-	}
-}
 
 type PalFXDef struct {
 	time      int32
@@ -537,44 +519,14 @@ func (s *Sprite) SetPxl(px []byte) {
 	}
 	sys.mainThreadTask <- func() {
 		s.Tex = newTexture()
-		gl.BindTexture(gl.TEXTURE_2D, s.Tex.handle)
-		gl.PixelStorei(gl.UNPACK_ALIGNMENT, 1)
-		gl.TexImage2D(gl.TEXTURE_2D, 0,
-			int(s.Size[0]), int(s.Size[1]),
-			gl.LUMINANCE, gl.UNSIGNED_BYTE,
-			px,
-		)
-		gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST)
-		gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST)
-		gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE)
-		gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE)
+		s.Tex.SetData(int32(s.Size[0]), int32(s.Size[1]), 8, false, px)
 	}
 }
 
 func (s *Sprite) SetRaw(data []byte, sprWidth int32, sprHeight int32, sprDepth int32) {
-	// TODO: Check why ths channel operation uses too much memory.
 	sys.mainThreadTask <- func() {
 		s.Tex = newTexture()
-		gl.BindTexture(gl.TEXTURE_2D, s.Tex.handle)
-		gl.PixelStorei(gl.UNPACK_ALIGNMENT, 1)
-		if sprDepth == 32 {
-			gl.TexImage2D(gl.TEXTURE_2D, 0,
-				int(sprWidth), int(sprHeight),
-				gl.RGBA, gl.UNSIGNED_BYTE, data)
-		} else {
-			gl.TexImage2D(gl.TEXTURE_2D, 0,
-				int(sprWidth), int(sprHeight),
-				gl.RGB, gl.UNSIGNED_BYTE, data)
-		}
-		if sys.pngFilter {
-			gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
-			gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
-		} else {
-			gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST)
-			gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST)
-		}
-		gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE)
-		gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE)
+		s.Tex.SetData(sprWidth, sprHeight, sprDepth, sys.pngFilter, data)
 	}
 }
 
@@ -1047,14 +999,10 @@ func (s *Sprite) glDraw(pal []uint32, mask int32, x, y float32, tile *[4]int32,
 		if hasPalette {
 			gl.BindTexture(gl.TEXTURE_2D, paltex.handle)
 		} else {
-			// Generate and cache palette texture
+			// Generate, bind and cache palette texture
 			s.PalTex = newTexture()
-			gl.BindTexture(gl.TEXTURE_2D, s.PalTex.handle)
-			gl.PixelStorei(gl.UNPACK_ALIGNMENT, 1)
-			gl.TexImage2D(gl.TEXTURE_2D, 0, 256, 1, gl.RGBA, gl.UNSIGNED_BYTE,
+			s.PalTex.SetData(256, 1, 32, false,
 				unsafe.Slice((*byte)(unsafe.Pointer(&pal[0])), len(pal) * 4))
-			gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST)
-			gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST)
 			tmp := append([]uint32{}, pal...)
 			s.paltemp = tmp
 		}
